third_party/blink/web_tests/virtual/mouse-selection-shadow-dom-disabled/external/wpt/shadow-dom/focus/text-selection-with-delegatesFocus-expected.txt

This is a testharness.js-based test.
[FAIL] shadow root has selectable text when focus is delegated
  assert_greater_than: expected a number greater than 0 but got 0
Harness: the test ran to completion.
